Expand the crisis response section.
Some words abuot how important procdures are for dealing with a crisis.
This is not the place to describe how to create a procedure (that looks
like it belongs in the Business Continuity Planning section under
Disaster Recovery), but it definitely is the place to describe how
useful they are for responding to disasters.